% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movienight.R
\name{movienight}
\alias{movienight}
\title{Recommends a movie}
\usage{
movienight(x)
}
\arguments{
\item{x}{Leave the argument empty}
}
\value{
A character vector
}
\description{
Recommends a movie
}
\examples{
movienight()
[1] "Up"
}
